requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.5
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
factory_boy==3.3.1
Faker==33.3.1
idna==3.10
inflection==0.5.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
